Fix: nitro structure initialization.
Simply calling memset() is not sufficient if the structure contains
pointers or floats, because the computer's internal representation of the
null pointer or 0.0 might not be all-0 bits.

See also <http://c-faq.com/malloc/calloc.html>.

The clever solution is to copy over a structure which the *complier*
initialized, and which is guaranteed to contain all the right bits.