Rename struct standard_header and struct OAM.
struct standard_header -> struct nitro

struct OAM -> struct OBJ

"standard_header" was never a great name, so it has been replaced with
something less generic and a bit more descriptive - "nitro" - which is
supposedly the name of Nintendo's DS SDK, and which loosely refers to
both the DS filesystem and the various file formats found therein.

(A better name might be "nitro_header", but "nitro" fits for
now and i plan to expand its scope in the near future.)

Some code had already been written which referred to instances of the
"OAM" struct as "obj", so this commit merely completes the transition.
"OAM" was really a synecdoche or misnomer, carried over from the llref
documentation, for "OBJ" - the former more properly referring to the
region of memory which the attributes of the OBJs occupy.

That said, the name "OBJ" is still a bit confusing due to its origin as
a truncated form of "object" (a much more generic concept). Perhaps
"VOBJ" (for Video OBJect) would be better. A thought for the future...