Drop .init and .initializer from nitro_type
The .initializer attribute was clever, but memset() works just as well
for our purposes. I'm dropping the .init method because it was never used
and it's bad to leave dead code lying around. It can be added back in
when we decide to do things properly.

I'm also dropping the "unsupported format" case because it was a shim
for formats we wanted to recognize but weren't yet implemented, and
those formats have all been implemented!

These changes simplify nitro_read_nocompressed() a bit.